# ai-server/src/api/swagger.yaml
openapi: 3.0.3
info:
  title: Multi-Universe Book Series Writing Assistant API
  version: 1.0.0
  description: |
    API for managing universes, books, chapters, documents, sections, and characters in a multi-universe book series writing assistant. All node types strictly follow canonical structure as defined in RAG_API_Caching_and_Parser_Consistency_Plan.md.

servers:
  - url: /api

paths:
  /characters:
    get:
      summary: List characters (optionally filtered by universe, book, chapter)
      tags: [Characters]
      parameters:
        - in: query
          name: universeId
          schema:
            type: string
          required: false
          description: Universe ID
        - in: query
          name: bookId
          schema:
            type: string
          required: false
          description: Book ID
        - in: query
          name: chapterId
          schema:
            type: string
          required: false
          description: Chapter ID
      responses:
        '200':
          description: Array of characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Character'
    post:
      summary: Create a new character
      tags: [Characters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Character'
      responses:
        '201':
          description: Character created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'

  /characters/{id}:
    put:
      summary: Update a character
      tags: [Characters]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Character'
      responses:
        '200':
          description: Character updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
    delete:
      summary: Delete a character
      tags: [Characters]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Character deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /api/tools:
    get:
      summary: List all registered tools
      description: Retrieve a list of all tools currently registered in the system.
      responses:
        '200':
          description: A list of tools.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
        '500':
          description: Internal server error.

  /api/tools/execute:
    post:
      summary: Execute a tool
      description: Execute a specific tool with the provided arguments.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                toolName:
                  type: string
                  description: The name of the tool to execute.
                args:
                  type: object
                  additionalProperties: true
                  description: The arguments to pass to the tool