# ai-server/src/routes/toolRoutes.ts
import { Router } from 'express';
import ToolOrchestratorController from '../controllers/toolOrchestratorController.js';

const router = Router();

/**
 * Route to execute a tool.
 * POST /api/tools/execute
 */
router.post('/execute', async (req, res) => {
  try {
    await ToolOrchestratorController.executeTool(req, res);
  } catch (error) {
    const errorMessage = error instanceof Error ? error.message : 'Unknown error';
    res.status(500).json({ success: false, error: errorMessage });
  }
});

/**
 * Route to list all registered tools.
 * GET /api/tools
 */
router.get('/', (req, res) => {
  try {
    const tools = require('../services/toolRegistryService').listTools();
    res.status(200).json({ success: true, tools });
  } catch (error) {
    console.error('Error listing tools:', error);
    const errorMessage = error instanceof Error ? error.message : 'Unknown error';
    res.status(500).json({ success: false, error: errorMessage });
  }
});

export default router;